Renamed features.js to feature-toggle.js
